+++ packages/paste-website/src/components/word-list/types.ts
@@ -1,0 +1,3 @@
+import wordsList from '../../data/word-list.json';
+
+export type WordListShape = typeof wordsList;
